Fix multiple Kotlin compilation errors
- TimerRing.kt: use this.size (DrawScope IntSize) instead of size param (Dp)
- SubjectsScreen.kt: add missing mutableIntStateOf import
- Type.kt: use GoogleFonts Font and import R
- StatisticsViewModel.kt: cast offset to Long for plusDays

+++ android/app/src/main/java/com/studyflow/app/viewmodel/StatisticsViewModel.kt
@@ -60,7 +60,7 @@ class StatisticsViewModel(application: Application) : AndroidViewModel(applicati
                 var weekTotal = 0
 
                 val weeklyData = (0..6).map { offset ->
-                    val date = monday.plusDays(offset)
+                    val date = monday.plusDays(offset.toLong())
                     val dateStr = date.format(formatter)
                     val minutes = dayMinutesMap[dateStr] ?: 0
                     weekTotal += minutes
